+++ src/styles/hero.scss
@@ -6,12 +6,12 @@ h1 {
 h2 {
   font-weight: 450;
   text-align: center;
+  font-size: 24px;
 }
 
 a {
   text-decoration-thickness: 1px;
-  font-size: 32px;
-
+  font-size: 24px;
   &:hover {
     color: #2661BA;
   }
